requirements.txt
streamlit>=1.20.0
spacy>=3.5.0
pydub>=0.25.1
pytest>=7.0.0
python-dotenv>=1.0.0
jsonschema>=4.0.0
moviepy>=1.0.3
imageio-ffmpeg>=0.4.7
Pillow>=9.0.0
redis>=4.0.0
rq>=1.1.0
